Misc Changes
1: Making Entity Update more robust

2: Updated Radar to have cull window for large map.
3: Updated plugin API onEnable/onDisable description
4: Updated plugin API onEnable function to pass in the information if the game is started or not.

5: Fixed the bug where ReadStdWString function throws an uncatched exception
6: Fixed iconpicker to not select an icon if user isn't hovering over iconpicker window.
7: Now old iconpicker will close when user wants to open a new iconpicker
8: Fixed the bug where Large Map Cull Window would close if clicked on any ImGui window

+++ GameHelper/GameProcess.cs
@@ -159,7 +159,7 @@ namespace GameHelper
                 // Have to check MainWindowHandle because
                 // sometime HasExited returns false even when game isn't running..
                 if (this.Information.HasExited ||
-                    this.Information.MainWindowHandle.ToInt32() <= 0x00)
+                    this.Information.MainWindowHandle.ToInt64() <= 0x00)
                 {
                     this.Close();
                     break;
